+++ tests/check_progress_nonmatchings.py
@@ -71,8 +71,8 @@ def getProgress(mapPath: Path, version: str, pathIndex: int=2) -> tuple[mapfile_
 
 cases: list[tuple[Path, str, mapfile_parser.ProgressStats]] = [
     (Path("tests/maps/gnuld/n64/drmario64.cn.map"),       "cn",  mapfile_parser.ProgressStats(undecompedSize=273028, decompedSize=199196)),
-    (Path("tests/maps/ld.lld/n64/drmario64.us.lld.map"),  "us",  mapfile_parser.ProgressStats(undecompedSize=170720, decompedSize=272764)),
-    (Path("tests/maps/gnuld/n64/drmario64.us.map"),       "us",  mapfile_parser.ProgressStats(undecompedSize=170720, decompedSize=272128)),
+    (Path("tests/maps/ld.lld/n64/drmario64.us.lld.map"),  "us",  mapfile_parser.ProgressStats(undecompedSize=170720, decompedSize=272860)),
+    (Path("tests/maps/gnuld/n64/drmario64.us.map"),       "us",  mapfile_parser.ProgressStats(undecompedSize=170720, decompedSize=272224)),
     (Path("tests/maps/gnuld/n64/puzzleleague64.usa.map"), "usa", mapfile_parser.ProgressStats(undecompedSize=263668, decompedSize=454604)),
 ]